pyproject.toml
[tool.poetry]
name = "lazarus-bot"
version = "1.0.0"
description = "A discord bot that automatically unarchives threads as soon as they're archived."
authors = ["Alex Oladele <dragid10@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8.2"
aiodns = "^3.0.0"
brotlipy = "^0.7.0"
cchardet = "^2.1.7"
loguru = "^0.6.0"
orjson = "^3.6.8"
py-cord = { extras = ["speed"], version = "^2.0.0-beta.7" }
python-decouple = "^3.6"
python-dotenv = "^0.20.0"
requests = "^2.27.1"
redis = "^4.2.2"


[tool.poetry.group.dev.dependencies]
icecream = "^2.1.2"


[tool.poetry.group.test.dependencies]
mockito = "^1.3.0"
pytest = "^7.1.1"
pytest-asyncio = "^0.18.3"
pytest-mockito = "0.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"